[CSL-1994] Move utility functions in Test.Pos to cardano-util
