+++ client/test/Test/Pos/Client/Txp/UtilSpec.hs
@@ -31,9 +31,10 @@ import           Pos.Crypto (RedeemSecretKey, SafeSigner, SecretKey, decodeHash,
                              redeemToPublic, toPublic)
 import           Pos.DB (gsAdoptedBVData)
 import           Pos.Txp (Utxo)
-import           Pos.Util.Arbitrary (nonrepeating)
+import           Pos.Util.QuickCheck.Arbitrary (nonrepeating)
+import           Pos.Util.QuickCheck.Property (stopProperty)
 import           Pos.Util.Util (leftToPanic)
-import           Test.Pos.Util (stopProperty, withDefConfigurations)
+import           Test.Pos.Configuration (withDefConfigurations)
 
 import           Test.Pos.Client.Txp.Mode (HasTxpConfigurations, TxpTestMode, TxpTestProperty,
                                            withBVData)
