+++ lib/cardano-sl.cabal
@@ -98,7 +98,6 @@ library
                         Test.Pos.Cbor.RefImpl
 
                         -- Util code for tests
-                        Test.Pos.Util
                         Test.Pos.Configuration
 
                         Pos.Util.Servant
