+++ lib/src/Test/Pos/Configuration.hs
@@ -1,20 +1,39 @@
+{-# LANGUAGE RankNTypes #-}
 -- | Default configuration for running tests.
 
 module Test.Pos.Configuration
        ( defaultTestConf
        , defaultTestGenesisSpec
        , defaultTestBlockVersionData
+
+       , HasStaticConfigurations
+       , withDefConfiguration
+       , withDefInfraConfiguration
+       , withDefNodeConfiguration
+       , withDefSscConfiguration
+       , withDefUpdateConfiguration
+       , withDefBlockConfiguration
+       , withDefDlgConfiguration
+       , withDefConfigurations
+       , withStaticConfigurations
        ) where
 
 import           Universum
 
 import qualified Data.Aeson as J
 
-import           Pos.Core (BlockVersionData)
+import           Pos.Block.Configuration (HasBlockConfiguration, withBlockConfiguration)
+import           Pos.Configuration (HasNodeConfiguration, withNodeConfiguration)
+import           Pos.Core (BlockVersionData, HasConfiguration, withGenesisSpec)
 import           Pos.Core.Configuration (CoreConfiguration (..), GenesisConfiguration (..))
 import           Pos.Core.Genesis (GenesisSpec (..))
-import           Pos.Launcher.Configuration (Configuration (..))
+import           Pos.Delegation (HasDlgConfiguration, withDlgConfiguration)
+import           Pos.Infra.Configuration (HasInfraConfiguration, withInfraConfiguration)
+import           Pos.Launcher.Configuration (Configuration (..), HasConfigurations)
+import           Pos.Ssc.Configuration (HasSscConfiguration, withSscConfiguration)
+import           Pos.Update.Configuration (HasUpdateConfiguration, withUpdateConfiguration)
 import           Pos.Util.Config (embedYamlConfigCT)
+
 
 -- | This configuration is embedded into binary and is used by default
 -- in tests.
@@ -33,3 +52,48 @@ defaultTestGenesisSpec =
 
 defaultTestBlockVersionData :: BlockVersionData
 defaultTestBlockVersionData = gsBlockVersionData defaultTestGenesisSpec
+
+-- | This constraint requires all configurations which are not
+-- always hardcoded in tests (currently).
+type HasStaticConfigurations =
+    ( HasInfraConfiguration
+    , HasUpdateConfiguration
+    , HasSscConfiguration
+    , HasBlockConfiguration
+    , HasNodeConfiguration
+    , HasDlgConfiguration
+    )
+
+withDefNodeConfiguration :: (HasNodeConfiguration => r) -> r
+withDefNodeConfiguration = withNodeConfiguration (ccNode defaultTestConf)
+
+withDefSscConfiguration :: (HasSscConfiguration => r) -> r
+withDefSscConfiguration = withSscConfiguration (ccSsc defaultTestConf)
+
+withDefUpdateConfiguration :: (HasUpdateConfiguration => r) -> r
+withDefUpdateConfiguration = withUpdateConfiguration (ccUpdate defaultTestConf)
+
+withDefInfraConfiguration :: (HasInfraConfiguration => r) -> r
+withDefInfraConfiguration = withInfraConfiguration (ccInfra defaultTestConf)
+
+withDefBlockConfiguration :: (HasBlockConfiguration => r) -> r
+withDefBlockConfiguration = withBlockConfiguration (ccBlock defaultTestConf)
+
+withDefDlgConfiguration :: (HasDlgConfiguration => r) -> r
+withDefDlgConfiguration = withDlgConfiguration (ccDlg defaultTestConf)
+
+withDefConfiguration :: (HasConfiguration => r) -> r
+withDefConfiguration = withGenesisSpec 0 (ccCore defaultTestConf)
+
+withStaticConfigurations :: (HasStaticConfigurations => r) -> r
+withStaticConfigurations patak =
+    withDefNodeConfiguration $
+    withDefSscConfiguration $
+    withDefUpdateConfiguration $
+    withDefBlockConfiguration $
+    withDefDlgConfiguration $
+    withDefInfraConfiguration patak
+
+withDefConfigurations :: (HasConfigurations => r) -> r
+withDefConfigurations bardaq =
+    withDefConfiguration $ withStaticConfigurations bardaq
