+++ lib/test/Test/Pos/Cbor/CborSpec.hs
@@ -39,7 +39,8 @@ import           Pos.Core.Common (ScriptVersion)
 import           Pos.Data.Attributes (Attributes (..), decodeAttributes, encodeAttributes)
 import qualified Test.Pos.Cbor.RefImpl as R
 import           Test.Pos.Helpers (binaryTest)
-import           Test.Pos.Util (withDefConfiguration, expectationError)
+import           Test.Pos.Configuration (withDefConfiguration)
+import           Pos.Util.QuickCheck.Property (expectationError)
 
 data User
     = Login { login :: String
