+++ lib/test/Test/Pos/Core/SeedSpec.hs
@@ -8,10 +8,11 @@ import           Universum
 
 import           Test.Hspec (Spec, describe)
 import           Test.Hspec.QuickCheck (prop)
-import           Test.Pos.Helpers (formsCommutativeMonoid)
+import           Test.Pos.Helpers ()
 import           Test.QuickCheck (Property, (.&&.))
 
 import           Pos.Core (SharedSeed)
+import           Pos.Util.QuickCheck.Property (formsCommutativeMonoid)
 
 spec :: Spec
 spec = describe "SharedSeed" $ do
