+++ lib/test/Test/Pos/Ssc/Identity/BinarySpec.hs
@@ -17,8 +17,8 @@ import           Pos.Core (StakeholderId)
 import qualified Pos.Core.Ssc as Ssc
 import qualified Pos.Ssc as Ssc
 
+import           Test.Pos.Configuration (withDefConfiguration, withDefInfraConfiguration)
 import           Test.Pos.Helpers (binaryTest, msgLenLimitedTest)
-import           Test.Pos.Util (withDefConfiguration, withDefInfraConfiguration)
 
 spec :: Spec
 spec = withDefInfraConfiguration $ withDefConfiguration $
