+++ lib/test/Test/Pos/Ssc/Identity/SafeCopySpec.hs
@@ -12,7 +12,7 @@ import qualified Pos.Core.Ssc as Ssc
 import           Pos.SafeCopy ()
 
 import           Test.Pos.Helpers (safeCopyTest)
-import           Test.Pos.Util (withDefConfiguration)
+import           Test.Pos.Configuration (withDefConfiguration)
 
 spec :: Spec
 spec = withDefConfiguration $ describe "Ssc" $ do
