+++ lib/test/Test/Pos/Ssc/VssCertDataSpec.hs
@@ -26,8 +26,9 @@ import           Pos.Ssc (SscGlobalState (..), VssCertData (..), delete, empty, 
                           insert, keys, lookup, member, rollbackSsc, runPureToss, setLastKnownSlot,
                           sgsVssCertificates)
 import           Pos.Util.Chrono (NewestFirst (..))
+import           Pos.Util.QuickCheck.Property (qcIsJust)
 
-import           Test.Pos.Util (qcIsJust, withDefConfiguration)
+import           Test.Pos.Configuration (withDefConfiguration)
 
 spec :: Spec
 spec = withDefConfiguration $ describe "Ssc.VssCertData" $ do
