+++ lib/test/Test/Pos/Txp/Identity/BinarySpec.hs
@@ -18,8 +18,8 @@ import qualified Pos.Txp as T
 import           Pos.Util (SmallGenerator)
 
 import           Test.Pos.Cbor.CborSpec (extensionProperty)
+import           Test.Pos.Configuration (withDefConfiguration, withDefInfraConfiguration)
 import           Test.Pos.Helpers (binaryTest, msgLenLimitedTest)
-import           Test.Pos.Util (withDefConfiguration, withDefInfraConfiguration)
 
 spec :: Spec
 spec = withDefInfraConfiguration $ withDefConfiguration $
