+++ lib/test/Test/Pos/Types/Identity/SafeCopySpec.hs
@@ -12,8 +12,8 @@ import qualified Pos.Core as Core
 import qualified Pos.Core.Txp as Txp
 import           Pos.SafeCopy ()
 
+import           Test.Pos.Configuration (withDefConfiguration)
 import           Test.Pos.Helpers (safeCopyTest)
-import           Test.Pos.Util (withDefConfiguration)
 
 spec :: Spec
 spec = withDefConfiguration $ describe "Types" $ do
