+++ lib/test/Test/Pos/UtilSpec.hs
@@ -14,10 +14,10 @@ import qualified Data.List.NonEmpty as NE
 import qualified GHC.Exts as IL (IsList (..))
 
 import           Pos.Util.Chrono (Chrono (..), NewestFirst (..), OldestFirst (..))
+import           Pos.Util.QuickCheck.Property ((.=.))
 
 import           Test.Hspec (Expectation, Spec, describe, shouldBe)
 import           Test.Hspec.QuickCheck (prop)
-import           Test.Pos.Helpers ((.=.))
 import           Test.QuickCheck (Arbitrary, Property)
 
 spec :: Spec
