+++ util/Pos/Util/QuickCheck/Property.hs
@@ -1,20 +1,6 @@
-{-# LANGUAGE GADTs      #-}
-{-# LANGUAGE RankNTypes #-}
-
-module Test.Pos.Util
-       ( HasStaticConfigurations
-       , withDefConfiguration
-       , withDefInfraConfiguration
-       , withDefNodeConfiguration
-       , withDefSscConfiguration
-       , withDefUpdateConfiguration
-       , withDefBlockConfiguration
-       , withDefDlgConfiguration
-       , withDefConfigurations
-       , withStaticConfigurations
-
-       -- * Various properties and predicates
-       , qcIsJust
+module Pos.Util.QuickCheck.Property
+       ( -- * Various properties and predicates
+         qcIsJust
        , qcIsNothing
        , qcIsLeft
        , qcIsRight
@@ -35,75 +21,26 @@ module Test.Pos.Util
        -- * Generators
        , splitWord
        , sumEquals
+
+       -- * Helpers
+       , (.=.)
+       , (>=.)
+       , shouldThrowException
+
+       -- * Semigroup/monoid laws
+       , formsSemigroup
+       , formsMonoid
+       , formsCommutativeMonoid
        ) where
 
 import           Universum
 
-import           Data.Tagged (Tagged (..))
-import           Test.QuickCheck (Arbitrary (arbitrary), Property, counterexample, property)
+import qualified Data.Semigroup as Semigroup
+import           Test.Hspec (Expectation, Selector, shouldThrow)
+import           Test.QuickCheck (Arbitrary, Property, counterexample, property, (.&&.), (===))
 import           Test.QuickCheck.Gen (Gen, choose)
 import           Test.QuickCheck.Monadic (PropertyM, pick, stop)
 import           Test.QuickCheck.Property (Result (..), failed)
-import qualified Test.Hspec as Hspec (Expectation)
-
-import           Pos.Block.Configuration (HasBlockConfiguration, withBlockConfiguration)
-import           Pos.Configuration (HasNodeConfiguration, withNodeConfiguration)
-import           Pos.Core (HasConfiguration, withGenesisSpec)
-import           Pos.Delegation (HasDlgConfiguration, withDlgConfiguration)
-import           Pos.Infra.Configuration (HasInfraConfiguration, withInfraConfiguration)
-import           Pos.Launcher.Configuration (Configuration (..), HasConfigurations)
-import           Pos.Ssc.Configuration (HasSscConfiguration, withSscConfiguration)
-import           Pos.Update.Configuration (HasUpdateConfiguration, withUpdateConfiguration)
-
-import           Test.Pos.Configuration (defaultTestConf)
-
--- | This constraint requires all configurations which are not
--- always hardcoded in tests (currently).
-type HasStaticConfigurations =
-    ( HasInfraConfiguration
-    , HasUpdateConfiguration
-    , HasSscConfiguration
-    , HasBlockConfiguration
-    , HasNodeConfiguration
-    , HasDlgConfiguration
-    )
-
-withDefNodeConfiguration :: (HasNodeConfiguration => r) -> r
-withDefNodeConfiguration = withNodeConfiguration (ccNode defaultTestConf)
-
-withDefSscConfiguration :: (HasSscConfiguration => r) -> r
-withDefSscConfiguration = withSscConfiguration (ccSsc defaultTestConf)
-
-withDefUpdateConfiguration :: (HasUpdateConfiguration => r) -> r
-withDefUpdateConfiguration = withUpdateConfiguration (ccUpdate defaultTestConf)
-
-withDefInfraConfiguration :: (HasInfraConfiguration => r) -> r
-withDefInfraConfiguration = withInfraConfiguration (ccInfra defaultTestConf)
-
-withDefBlockConfiguration :: (HasBlockConfiguration => r) -> r
-withDefBlockConfiguration = withBlockConfiguration (ccBlock defaultTestConf)
-
-withDefDlgConfiguration :: (HasDlgConfiguration => r) -> r
-withDefDlgConfiguration = withDlgConfiguration (ccDlg defaultTestConf)
-
-withDefConfiguration :: (HasConfiguration => r) -> r
-withDefConfiguration = withGenesisSpec 0 (ccCore defaultTestConf)
-
-withStaticConfigurations :: (HasStaticConfigurations => r) -> r
-withStaticConfigurations patak =
-    withDefNodeConfiguration $
-    withDefSscConfiguration $
-    withDefUpdateConfiguration $
-    withDefBlockConfiguration $
-    withDefDlgConfiguration $
-    withDefInfraConfiguration patak
-
-withDefConfigurations :: (HasConfigurations => r) -> r
-withDefConfigurations bardaq =
-    withDefConfiguration $ withStaticConfigurations bardaq
-
-instance Arbitrary a => Arbitrary (Tagged s a) where
-    arbitrary = Tagged <$> arbitrary
 
 ----------------------------------------------------------------------------
 -- Various properties and predicates
@@ -208,5 +145,69 @@ sumEquals maxEl restSum = do
     el <- choose (1, min maxEl restSum)
     (el:) <$> sumEquals maxEl (restSum - el)
 
-expectationError :: Text -> Hspec.Expectation
+expectationError :: Text -> Expectation
 expectationError = fail . toString
+
+----------------------------------------------------------------------------
+-- Monoid/Semigroup laws
+----------------------------------------------------------------------------
+
+isAssociative :: (Show m, Eq m, Semigroup m) => m -> m -> m -> Property
+isAssociative m1 m2 m3 =
+    let assoc1 = (m1 Semigroup.<> m2) Semigroup.<> m3
+        assoc2 = m1 Semigroup.<> (m2 Semigroup.<> m3)
+    in assoc1 === assoc2
+
+formsSemigroup :: (Show m, Eq m, Semigroup m) => m -> m -> m -> Property
+formsSemigroup = isAssociative
+
+hasIdentity :: (Eq m, Semigroup m, Monoid m) => m -> Property
+hasIdentity m =
+    let id1 = mempty Semigroup.<> m
+        id2 = m Semigroup.<> mempty
+    in (m == id1) .&&. (m == id2)
+
+formsMonoid :: (Show m, Eq m, Semigroup m, Monoid m) => m -> m -> m -> Property
+formsMonoid m1 m2 m3 =
+    (formsSemigroup m1 m2 m3) .&&. (hasIdentity m1)
+
+isCommutative :: (Show m, Eq m, Semigroup m, Monoid m) => m -> m -> Property
+isCommutative m1 m2 =
+    let comm1 = m1 <> m2
+        comm2 = m2 <> m1
+    in comm1 === comm2
+
+formsCommutativeMonoid :: (Show m, Eq m, Semigroup m, Monoid m) => m -> m -> m -> Property
+formsCommutativeMonoid m1 m2 m3 =
+    (formsMonoid m1 m2 m3) .&&. (isCommutative m1 m2)
+
+----------------------------------------------------------------------------
+-- Helpers
+----------------------------------------------------------------------------
+
+-- | Extensional equality combinator. Useful to express function properties as functional
+-- equations.
+(.=.) :: (Eq b, Show b, Arbitrary a) => (a -> b) -> (a -> b) -> a -> Property
+(.=.) f g a = f a === g a
+
+infixr 5 .=.
+
+-- | Monadic extensional equality combinator.
+(>=.)
+    :: (Show (m b), Arbitrary a, Monad m, Eq (m b))
+    => (a -> m b)
+    -> (a -> m b)
+    -> a
+    -> Property
+(>=.) f g a = f a === g a
+
+infixr 5 >=.
+
+shouldThrowException
+    :: (Show a, Eq a, Exception e)
+    => (a -> b)
+    -> Selector e
+    -> a
+    -> Expectation
+shouldThrowException action exception arg =
+    (return $! action arg) `shouldThrow` exception
