+++ wallet-new/test/unit/UTxO/Translate.hs
@@ -39,7 +39,7 @@ import qualified UTxO.Verify as Verify
   configuration.yaml. It is specified by a 'GenesisSpec'.
 -------------------------------------------------------------------------------}
 
-import Test.Pos.Util (
+import Test.Pos.Configuration (
     withDefConfiguration
   , withDefUpdateConfiguration
   )
