+++ wallet/test/Test/Pos/Wallet/Web/AddressSpec.hs
@@ -19,13 +19,14 @@ import           Pos.Crypto (PassPhrase)
 import           Pos.Launcher (HasConfigurations)
 import           Pos.Util.CompileInfo (HasCompileInfo, withCompileInfo)
 
+import           Pos.Util.QuickCheck.Property (assertProperty, expectedOne)
 import           Pos.Wallet.Web.Account (GenSeed (..), genUniqueAddress)
 import           Pos.Wallet.Web.ClientTypes (AccountId, CAccountInit (..), caId, cwamId)
 import           Pos.Wallet.Web.Error (WalletError (..))
 import           Pos.Wallet.Web.Methods.Logic (newAccount)
 import           Pos.Wallet.Web.State (getWalletAddresses)
 import           Pos.Wallet.Web.Util (decodeCTypeOrFail)
-import           Test.Pos.Util (assertProperty, expectedOne, withDefConfigurations)
+import           Test.Pos.Configuration (withDefConfigurations)
 import           Test.Pos.Wallet.Web.Mode (WalletProperty)
 import           Test.Pos.Wallet.Web.Util (importSingleWallet, mostlyEmptyPassphrases)
