+++ wallet/test/Test/Pos/Wallet/Web/Util.hs
@@ -36,8 +36,8 @@ import           Pos.Client.Txp.Balances (getBalance)
 import           Pos.Core (Address, BlockCount, Coin, HasConfiguration, genesisSecretsPoor,
                            headerHashG)
 import           Pos.Core.Block (blockHeader)
-import           Pos.Core.Genesis (poorSecretToEncKey)
 import           Pos.Core.Common (IsBootstrapEraAddr (..), deriveLvl2KeyPair)
+import           Pos.Core.Genesis (poorSecretToEncKey)
 import           Pos.Core.Txp (TxIn, TxOut (..), TxOutAux (..))
 import           Pos.Crypto (EncryptedSecretKey, PassPhrase, ShouldCheckPassphrase (..),
                              emptyPassphrase, firstHardened)
@@ -48,13 +48,13 @@ import           Pos.Txp.Toil (Utxo)
 import           Pos.Util (HasLens (..), _neLast)
 import           Pos.Util.Chrono (OldestFirst (..))
 import           Pos.Util.CompileInfo (HasCompileInfo)
+import           Pos.Util.QuickCheck.Property (assertProperty, maybeStopProperty)
 import           Pos.Util.Servant (encodeCType)
 import           Pos.Util.UserSecret (mkGenesisWalletUserSecret)
 import           Pos.Wallet.Web.ClientTypes (Addr, CId, Wal, encToCId)
 import           Pos.Wallet.Web.Methods.Restore (importWalletDo)
 
 import           Test.Pos.Block.Logic.Util (EnableTxPayload, InplaceDB, genBlockGenParams)
-import           Test.Pos.Util (assertProperty, maybeStopProperty)
 import           Test.Pos.Wallet.Web.Mode (WalletProperty)
 
 ----------------------------------------------------------------------------
